Add Typing rule category
Contains contents of 'best_practices/kinds.rs', split across
multiple files.
