+++ src/rules/best_practices/mod.rs
@@ -1,4 +1,3 @@
 pub mod filesystem;
 pub mod implicit_none;
-pub mod kinds;
 pub mod modules_and_programs;
