+++ src/rules/typing/mod.rs
@@ -1,0 +1,3 @@
+pub mod literal_kinds;
+pub mod real_precision;
+pub mod star_kinds;
